python/gcs/requirements.txt
backcall==0.2.0
colorama==0.4.4
decorator==5.1.0
et-xmlfile==1.1.0
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
meson==0.59.2
ninja==1.10.2.2
openpyxl==3.0.9
ordereddict==1.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2==2.9.3
Pygments==2.10.0
PyYAML==6.0
traitlets==5.1.0
wcwidth==0.2.5
xmlplain==1.6.0
